Add "create-new-book" modal dialog
Created a modal dialog that is opened whenever the user clicks the
"Add new book" button. It has been given some basic styling.
The modal contains the a form for the user to complete with the book's
information (title, author, number of pages, and if it was read).
Two buttons are created, "Create" will be used in the future to gather
the user's data and store it in the script. "Close" triggers an event
that closes the modal dialog.