Add book object/div creation functionality
Function book is an Object Constructor containing book properties:
title, author, pages, and read.

Array myLibrary is created. It stores the book objects created by the
user.

submitBookButton() function triggers whenever the user clicks the create
button within the "Add new book" dialog. The following functionalities
have been added to the function (more to come):

--> push a new book object to the myLibrary Array. The object contains
the properties according to the users form completion.

--> close the dialog

--> loop through myLibraryArray (with stored objects). For each object
stored, create a div, give it an id equal to the object array's index,
give it "grid-item" class so it can be styled through css, and append it.
This will create as many divs as objects the user creates.