Add text content to book divs
The for loop in function "submitBookButton" now creates new html elements
and appends them to the book divs. Each element is added text content
to make sure each div correctly displays its associated book properties:
title, author, number of pages, and if it has been read or not. An img
is also added to be used in the future as the book div delete button.

The addNewBookButton event listener now resets as well the form values,
so they opening the dialog always shows a clean sheet.