Refactor script.js previous code and habilitate book card deletion
Refactored previous code in script.js file to minimize as much as I
could the use of global variables as well as simplify the ones already
created:

- Now only array "myLibrary" and variable "bookCount" are created in the
global scope.

- "bookCount" is used to keep track of the current number of books. It
increments when a new book is created and decrements when it is deleted.

- All DOM reference variables that were previously created in the global
scope are now created inside of the corresponding functions or variables
, so that the code is cleanear and the possibility of crossing variable
names is low.

- "Book" function constructor is now a factory function instead. It uses
closures to create new book objects from it. Now it doesn't return the
data as the user inputs it. Instead it modifies its value and transforms
it to look as it is expected to be displayed on the book card. This saves
modifying it later with the creation of each book card.

- Creation of a new book object and push to "myLibrary" array is now
executed by "pushBookToArray()" closure, via function "createNewBookCard
()", which in turn is executed on variable "executeNewBookCreation".

- Function "createNewBookCard()" doesn't loop through the "myLibrary"
array anymore to gather the data and create the needed book card. Instead,
it saves the "pushBookToArray()" closure into "bookCard" variable, and
uses the later to extract the required values. Thus pushing the new book
object into the array and attaching the data to the div created altogether.

- Variable "executeNewBookCreation" is created to contain the event
listener that triggers the creation of a new book object in "myLibrary"
array and new book card in the display (through the execution of
"createNewBookCard()). It also increments "bookCount" variable.

Habilitated the book card deletion functionality:

- Function "deleteBookFromArray()" is passed down the id of the targeted
book card (a number which reflects the object position in the myLibrary
array) and splices the related book object from "myLibrary". It also
loops through the book cards were created after the one deleted, to
change its id's according to the new position of their related object in
the "myLibrary" array. This ensures that any other deletions will still
target the correct book card and array object.

- Function "deleteBookCard()" takes the user-targeted book card's id and
deletes the associated book card from the display.

- Function "executeBookDeletion()" triggers when the user clicks the
delete button of a book card (event assigned to each div on the
"createNewBookCard()" function). It opens a new dialog modal with two
buttons to confirm or cancel deletion. If user confirms, it executes
functions "deleteBookFromArray" and "deleteBookCard", effectively
deleting the targeted book card from the display and object from
"myLibrary".