Add change-read-status on click functionality
Created function "chageReadStatus". It triggers on a click event on the
read button (the event is added to the html through createNewBookCard()
function). Depending on the button's current textContent ("Read" or not)
it changes the button's textContent, its background colour, and the
associated array book's read value to the opposite, and viceversa.