Add stats functionality
Added a new modal dialog for the stats functionality.

Created variables "openStatsDialog" and "closeStatsDialog" to open and
close the stats dialog on click of the correct button.

Created function "updateStats". It creates and updates some variables
showing statistics through looping the "myLibrary" array. Then it
attached testContent showcasing the different stats on the different
dialog divs. If there are no books created. A message appears to justify
the lack of stats.