Create validation functionality on empty required camps in the form.
Integrated a simple validation functionality in function executeNewBook
Creation in order to not submit the book unless all form required
camps have been filled. An if statement checks if the value of any
required items is false, alerting a message to remind the user all
required camps must be filled when that is the case.