Add validity check for pages input. Style delete-book dialog
Added a pattern attribute to the pages input along with a RegEx that
only allows for an integer number value. --> Added the check functionality
to the executeNewBookCreation function by setting up another branch in
the if statment, alerting that pages camp requires an integer number
when it has a value but it doesn't fit the pattern attribute (used
the checkvalidity() method to accomplish that).

Added also border color styling for form inputs depending on the validity
or not of their value.

Styled the delete-book dialog.